add option to automatically tag after upload

+++ src/szurubooru_toolkit/saucenao.py
@@ -31,7 +31,7 @@ class SauceNao:
         self.konachan = Moebooru('konachan', config.konachan['user'], config.konachan['password'])
 
     @sync
-    async def get_metadata(self, post_url: str, szuru_public: bool, tmp_path: str):
+    async def get_metadata(self, post_url: str, szuru_public: bool, tmp_path: str, tmp_media_path: str = None):
         """
         Scrape and collect tags, sources, and ratings from popular imageboards
         Simply put, it's a wrapper for get_result() and scrape_<image_board>()
@@ -57,7 +57,7 @@ class SauceNao:
         limit_short = 1
         limit_long = 10
 
-        response = await self.get_result(post_url, szuru_public, tmp_path)
+        response = await self.get_result(post_url, szuru_public, tmp_path, tmp_media_path)
 
         # Sometimes multiple results from the same Booru are found.
         # Results are sorted by their similiarity (highest first).
@@ -171,7 +171,7 @@ class SauceNao:
 
         return tags, source, rating, limit_short, limit_long
 
-    async def get_result(self, post_url: str, szuru_public: bool, tmp_path: str):
+    async def get_result(self, post_url: str, szuru_public: bool, tmp_path: str, tmp_media_path: str = None):
         """
         If szurubooru is public, let SauceNAO fetch the image from supplied URL.
         If not not, download the image to our temporary path and upload it to SauceNAO.
@@ -198,8 +198,12 @@ class SauceNao:
         else:
             for _ in range(1, 12):
                 try:
-                    filename = post_url.split('/')[-1]
-                    tmp_file = urllib.request.urlretrieve(post_url, Path(tmp_path) / filename)[0]
+                    if not tmp_media_path:
+                        filename = post_url.split('/')[-1]
+                        tmp_file = urllib.request.urlretrieve(post_url, Path(tmp_path) / filename)[0]
+                    else:
+                        tmp_file = tmp_media_path
+
                     logger.debug(f'Trying to get result from tmp_file: {tmp_file}')
 
                     # Resize images larger than 2MB to reduce load on servers
@@ -211,9 +215,10 @@ class SauceNao:
                     response = await self.pysaucenao.from_file(tmp_file)
                     logger.debug(f'Received response {response}')
 
-                    # Remove temporary image
-                    if os.path.exists(tmp_file):
-                        os.remove(tmp_file)
+                    # Remove temporary image if the script was not called from upload-media
+                    if not tmp_media_path:
+                        if os.path.exists(tmp_file):
+                            os.remove(tmp_file)
 
                     break
                 except (ContentTypeError, TimeoutError):
